added explicit pathing based on os.path for static files

+++ utilities.py
@@ -23,13 +23,14 @@ def delete_files(file_paths):
     if file_paths:
      
         for file_path in file_paths:
-            try:
-                os.remove(f"./image_temp/{file_path}")
-                print(f"Remove : ./image_temp/{file_path}")
-            except OSError as e:
-                print(f"Error: {file_path} : {e.strerror}")
+            if file_path != ".keep":
+                try:
+                    os.remove(f"./image_temp/{file_path}")
+                    print(f"Remove : ./image_temp/{file_path}")
+                except OSError as e:
+                    print(f"Error: {file_path} : {e.strerror}")
         file_paths.clear()
-        
+            
 
 def open_image_from_url(image_url):
     response = requests.get(image_url)
